refactor, fall back to logger if not tty
