+++ cli/npm/registry.rs
@@ -21,7 +21,7 @@ use serde::Serialize;
 use crate::file_fetcher::CacheSetting;
 use crate::fs_util;
 use crate::http_cache::CACHE_PERM;
-use crate::proc_state::ProgressBar;
+use crate::progress_bar::ProgressBar;
 
 use super::cache::NpmCache;
 use super::semver::NpmVersionReq;
@@ -307,8 +307,7 @@ impl NpmRegistryApi {
     let package_url = self.get_package_url(name);
 
     {
-      let msg = format!("{} {}", colors::green("Download"), package_url);
-      self.progress_bar.update(msg);
+      self.progress_bar.update(package_url.as_str());
     }
 
     let response = match reqwest::get(package_url).await {
